bot/requirements.txt
python-binance
redis
pandas
pydantic